# docs/swagger.yaml
definitions:
...
  domain.Product:
    properties:
      description:
        type: string
      expiration_rate:
        type: integer
      freezing_rate:
        type: integer
      height:
        type: number
      id:
        type: integer
      length:
        type: number
      netweight:
        type: number
      product_code:
        type: string
      product_type_id:
        type: integer
      recommended_freezing_temperature:
        type: number
      seller_id:
        type: integer
      width:
        type: number
    required:
    - description
    - expiration_rate
    - freezing_rate
    - height
    - length
    - netweight
    - product_code
    - product_type_id
    - recommended_freezing_temperature
    - width
    type: object
...
  domain.ProductRequest:
    properties:
      description:
        type: string
      expiration_rate:
        type: integer
      freezing_rate:
        type: integer
      height:
        type: number
      length:
        type: number
      netweight:
        type: number
      product_code:
        type: string
      product_type_id:
        type: integer
      recommended_freezing_temperature:
        type: number
      seller_id:
        type: integer
      width:
        type: number
    required:
    - description
    - expiration_rate
    - freezing_rate
    - height
    - length
    - netweight
    - product_code
    - product_type_id
    - recommended_freezing_temperature
    - width
    type: object
  domain.ProductType:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
  domain.ProductTypeRequest:
    properties:
      name:
        type: string
    required:
    - name
    type: object
...
  domain.Report:
    properties:
      description:
        type: string
      product_id:
        type: integer
      records_count:
        type: integer
    type: object
...
  web.errorResponse:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
  web.response:
    properties:
      data: {}
    type: object
info:
  contact: {}
  description: This API was created as part of Mercado Libre's Go backend bootcamp
    for training purposes.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://developers.mercadolibre.com.ar/es-ar/terminos-y-condiciones
  title: MeLi Bootcamp API
  version: "1.0"
paths:
...
  /api/v1/products:
    get:
      description: Returns a list of all products
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/web.response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/domain.Product'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.errorResponse'
      summary: List products
      tags:
      - Products
  /api/v1/products/:
    post:
      consumes:
      - application/json
      description: Creates and returns a single product
      parameters:
      - description: Product parameters
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.ProductRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/web.response'
            - properties:
                data:
                  $ref: '#/definitions/domain.Product'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.errorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/web.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.errorResponse'
      summary: Create product
      tags:
      - Products
  /api/v1/products/{id}:
    delete:
      description: Deletes the product specified by URL id parameter.
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            $ref: '#/definitions/web.response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/web.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.errorResponse'
      summary: Delete product
      tags:
      - Products
    get:
      description: Returns a single product specified by its ID passed as a URL parameter
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/web.response'
            - properties:
                data:
                  $ref: '#/definitions/domain.Product'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/web.errorResponse'
      summary: Get product by ID
      tags:
      - Products
    patch:
      consumes:
      - application/json
      description: 'Updates the product specified by URL id parameter with fields
        passed by request body. All object fields are optional: only the given fields
        will be updated. Returns updated object.'
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      - description: Product parameters
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.ProductRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/web.response'
            - properties:
                data:
                  $ref: '#/definitions/domain.Product'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/web.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.errorResponse'
      summary: Update product
      tags:
      - Products
  /api/v1/products/reportRecords:
    get:
      description: Given a product id as a query, it will return the amount of product
        records for that given product. If given no id, it will return the amount
        of product records for all products.
      parameters:
      - description: Product ID
        in: query
        name: id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/web.response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/domain.Report'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/web.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.errorResponse'
      summary: Get product record report
      tags:
      - Products
  /api/v1/products/type:
    post:
      consumes:
      - application/json
      description: Given a name, creates a product type with that name
      parameters:
      - description: Product type
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.ProductTypeRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/web.response'
            - properties:
                data:
                  $ref: '#/definitions/domain.ProductType'
              type: object
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/web.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.errorResponse'
      summary: Create product type
      tags:
      - Products
...
securityDefinitions:
  ApiKeyAuth:
    description: start with Bearer
    in: header
    name: x-tiger-token
    type: apiKey
swagger: "2.0"